refactor: 코드 리뷰 반영 및 api 모듈로 이동
