+++ backend/techpick-api/src/main/java/techpick/api/infrastructure/lock/LockException.java
@@ -1,0 +1,8 @@
+package techpick.api.infrastructure.lock;
+
+public class LockException extends RuntimeException {
+
+	public LockException(String message) {
+		super(message);
+	}
+}
