+++ backend/techpick-api/src/main/java/techpick/api/infrastructure/lock/LockProvider.java
@@ -1,4 +1,4 @@
-package techpick.core.util;
+package techpick.api.infrastructure.lock;
 
 /**
  * @author sangwon
